feat(rolereact): split the code in multiple files
